+++ src/Kupo/Data/Health.hs
@@ -24,12 +24,17 @@ import Data.ByteString.Builder
 import Data.ByteString.Builder.Scientific
     ( formatScientificBuilder
     )
+import Data.Ratio
+    ( (%)
+    )
 import Data.Scientific
     ( FPFormat (Fixed)
     , Scientific
     )
 import Data.Time
-    ( UTCTime
+    ( NominalDiffTime
+    , UTCTime
+    , diffUTCTime
     )
 import Kupo.Data.Cardano
     ( Point
@@ -40,6 +45,7 @@ import Kupo.Data.Cardano
 import Kupo.Data.Configuration
     ( DeferIndexesInstallation (..)
     , NetworkParameters (..)
+    , SystemStart (..)
     )
 import Kupo.Version
     ( version
@@ -64,9 +70,13 @@ import System.Metrics.Prometheus.MetricId
 import System.Metrics.Prometheus.Registry
     ( RegistrySample (..)
     )
+import Ouroboros.Network.Magic
+    ( NetworkMagic (..)
+    )
 
 import qualified Data.Aeson.Encoding as Json
 import qualified Data.Map as Map
+import qualified Data.Scientific as Scientific
 
 -- | Information about the overall state of the application.
 data Health = Health
@@ -82,12 +92,16 @@ data Health = Health
         -- ^ Some useful server configuration
     } deriving stock (Generic, Eq, Show)
 
-data SerialisableHealth = SerialisableHealth (Maybe NetworkParameters) Health
+data SerialisableHealth = SerialisableHealth
+    { health :: Health
+    , networkParameters :: Maybe NetworkParameters
+    , currentTime :: UTCTime
+    }
 
 instance ToJSON SerialisableHealth where
     toJSON =
         error "'toJSON' called on 'Health'. This should never happen. Use 'toEncoding' instead."
-    toEncoding (SerialisableHealth _ Health{..}) = Json.pairs $ mconcat
+    toEncoding (SerialisableHealth Health{..} optNetworkParameters now) = Json.pairs $ mconcat
         [ Json.pair
             "connection_status"
             (toEncoding connectionStatus)
@@ -98,6 +112,17 @@ instance ToJSON SerialisableHealth where
             "most_recent_node_tip"
             (maybe Json.null_ slotNoToJson mostRecentNodeTip)
         , Json.pair
+            "seconds_since_last_block"
+            (maybe Json.null_ (nominalDiffTimeToJson . max 0 . diffUTCTime now) mostRecentClockTick)
+        , Json.pair
+            "network_synchronization"
+            (maybe Json.null_ toEncoding
+                (liftA2 (mkNetworkSynchronization now)
+                    optNetworkParameters
+                    (getPointSlotNo <$> mostRecentCheckpoint)
+                )
+            )
+         , Json.pair
             "configuration"
             (Json.pairs $ mconcat
                 [ Json.pair
@@ -109,6 +134,16 @@ instance ToJSON SerialisableHealth where
             "version"
             (toEncoding version)
         ]
+
+-- | Encode a 'NominalDiffTime' as an integer representing number of seconds.
+nominalDiffTimeToJson :: NominalDiffTime -> Json.Encoding
+nominalDiffTimeToJson =
+    Json.integer . nominalDiffTimeToInteger
+
+-- | Encode a 'NominalDiffTime' as an integer representing number of seconds.
+nominalDiffTimeToInteger :: NominalDiffTime -> Integer
+nominalDiffTimeToInteger =
+    (`div` 10^(12::Integer)) . toInteger . fromEnum
 
 emptyHealth :: Health
 emptyHealth = Health
@@ -130,8 +165,73 @@ instance ToJSON ConnectionStatus where
         Connected -> Json.text "connected"
         Disconnected -> Json.text "disconnected"
 
-mkPrometheusMetrics :: Health -> Builder
-mkPrometheusMetrics Health{..} =
+-- | Captures how far is our underlying node from the network, in percentage.
+newtype NetworkSynchronization = NetworkSynchronization Scientific
+    deriving stock (Generic, Eq, Ord, Show)
+
+instance ToJSON NetworkSynchronization where
+    toJSON _ =
+        error "'toJSON' called on 'NetworkSynchronization'. This should never happen. Use 'toEncoding' instead."
+    toEncoding (NetworkSynchronization s) =
+        -- NOTE: Using a specific encoder here to avoid turning the value into
+        -- scientific notation. Indeed, for small decimals values, aeson
+        -- automatically turn the representation into scientific notation with
+        -- exponent. While this is useful and harmless in many cases, it makes
+        -- consuming this value a bit harder from scripts. Since we know (by
+        -- construction) that the network value will necessarily have a maximum
+        -- of 5 decimals, we encode it as a number with a fixed number (=5) of
+        -- decimals.
+        --
+        -- >>> encode (NetworkSynchronization 1.0)
+        -- 1.00000
+        --
+        -- >>> encode (NetworkSynchronization 1.4e-3)
+        -- 0.00140
+        --
+        -- etc...
+        Json.unsafeToEncoding (formatScientificBuilder Fixed (Just 5) s)
+
+mkNetworkSynchronization
+    :: UTCTime -- Current Time
+    -> NetworkParameters
+    -> SlotNo -- Last known tip slot
+    -> NetworkSynchronization
+mkNetworkSynchronization now NetworkParameters { systemStart = SystemStart systemStart, networkMagic } (SlotNo lastKnownTip) =
+    let
+        tip = fromIntegral @_ @Integer lastKnownTip
+
+        -- Number of slots that happened in the Byron era, based on the Network Magic
+        offset = case networkMagic of
+            NetworkMagic 764824073 -> 4492799
+            NetworkMagic 1 -> 84242
+            NetworkMagic 2 -> 0
+            _ -> 0
+
+        byronSlotLength = 20
+
+        preByron
+            | tip >= offset = offset * byronSlotLength
+            | otherwise = tip * byronSlotLength
+
+        postByron
+            | tip >= offset = tip - offset
+            | otherwise = 0
+
+        num = preByron + postByron
+        den = round @_ @Integer (now `diffUTCTime` systemStart) + offset * (byronSlotLength - 1)
+
+        tolerance = 120
+        p = 100000
+    in
+        if abs (num - den) <= tolerance then
+            NetworkSynchronization 1
+        else
+            NetworkSynchronization
+                $ Scientific.unsafeFromRational
+                $ min 1 (((num * p) `div` den) % p)
+
+mkPrometheusMetrics :: UTCTime -> Maybe NetworkParameters -> Health -> Builder
+mkPrometheusMetrics now optNetworkParameters Health{..} =
     prometheusMetrics
     & Map.fromList
     & Map.mapKeys (\k -> (MetricId (makeName $ "kupo_" <> k) (Labels mempty)))
@@ -143,6 +243,12 @@ mkPrometheusMetrics Health{..} =
 
     mkCounter :: Int -> MetricSample
     mkCounter = CounterMetricSample . CounterSample
+
+    networkSynchronization :: Maybe NetworkSynchronization
+    networkSynchronization = liftA2
+        (mkNetworkSynchronization now)
+        optNetworkParameters
+        (getPointSlotNo <$> mostRecentCheckpoint)
 
     prometheusMetrics :: [(Text, MetricSample)]
     prometheusMetrics = mconcat
@@ -161,6 +267,16 @@ mkPrometheusMetrics Health{..} =
         , [ ( "most_recent_node_tip"
             , mkCounter $ fromEnum $ unSlotNo s
             ) | Just s <- [mostRecentNodeTip]
+          ]
+
+        , [ ( "seconds_since_last_block"
+            , mkGauge $ fromIntegral $ nominalDiffTimeToInteger $ max 0 $ diffUTCTime now t
+            ) | Just t <- [mostRecentClockTick]
+          ]
+
+        , [ ( "network_synchronization"
+            , mkGauge $ Scientific.toRealFloat s
+            ) | Just (NetworkSynchronization s) <- [networkSynchronization]
           ]
 
         , [ ( "configuration_indexes"